requirements.txt
dnspython
requests
beautifulsoup4
python-nmap
python-whois
DateTime